% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diseaseSpread.R
\name{diseaseSpread}
\alias{diseaseSpread}
\title{A disease spread function}
\usage{
diseaseSpread(mat, startnode, p, timesteps, sims)
}
\arguments{
\item{mat}{adjacency matrix of 0 and 1s representing potential links which will transmit infection from an infected to an uninfected node at probability described in p}

\item{startnode}{the node where infection will start}

\item{p}{probability of infection}

\item{timesteps}{number of timesteps per simulation}

\item{sims}{number of simulations}

\item{nodenum}{total number of nodes in the network}
}
\description{
This function allows you to simulate disease spread through a network.
}
\examples{
diseasespread()
}
\keyword{network}
\keyword{spread}
