% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DiseaseSpreadPWR.R
\name{DiseaseSpreadPWR}
\alias{DiseaseSpreadPWR}
\title{A disease spread function}
\usage{
DiseaseSpreadPWR(transmat, d, startnode, p, Beta)
}
\arguments{
\item{transmat}{your transaction matrix which is a full network matrix of 1 and 0s}

\item{d}{matrix of geographic distances between all nodes, scale does not matter}

\item{startnode}{single node number that introduces the pathogen into the network}

\item{p}{probability of infection on a link in each timestep}

\item{Beta}{spread parameter beta that will be used in the inverse power law model to model link formation between locations- higher number means fewer links}
}
\description{
This function allows you to simulate disease spread through a network that combines two types of data- observed transaction data and modeled spread data. Spread modeled using the inverse power law model.
}
\examples{
DiseaseSpreadPWR()
}
\keyword{network}
\keyword{spread}
